phase3: split simulation functions into src/simulation.cpp
Extracted Evolution, UpdateSeeds, UpdateField, FillSeed, RecruitTree,
TriggerTreefall, TriggerTreefallSecondary from troll.cpp (337–1073)
into src/simulation.cpp. troll.cpp now contains only main() (~338 lines).
